Improve the annotateByGene and annotateByHgvs

+++ src/main/java/org/mozi/varann/data/impl/annotation/VariantContextToEffectRecordConverter.java
@@ -30,7 +30,7 @@ public class VariantContextToEffectRecordConverter implements VariantContextToRe
         builder.setRef(vc.getReference().getBaseString());
         for (Allele all : vc.getAlternateAlleles()) {
             builder.getAlt().add(all.getBaseString());
-            GenomeVariant gv = new GenomeVariant(new GenomePosition(refDict, Strand.FWD, refDict.getContigNameToID().get(vc.getContig()), builder.getPos(), PositionType.ONE_BASED), builder.getRef(), all.getBaseString());
+            GenomeVariant gv = new GenomeVariant(new GenomePosition(refDict, Strand.FWD, refDict.getContigNameToID().get(vc.getContig()), (int)builder.getPos(), PositionType.ONE_BASED), builder.getRef(), all.getBaseString());
             builder.getHgvs().add(gv.toString());
         }
