+++ src/main/java/org/mozi/varann/data/impl/dbnsfp/DBNSFPRecordConverter.java
@@ -29,7 +29,7 @@ public class DBNSFPRecordConverter implements TSVToRecordConverter<DBNSFPRecord,
         builder.setPos(Integer.parseInt(record.get("hg19_pos(1-based)")));
         builder.setRef(record.get("ref"));
         builder.getAlt().add(record.get("alt"));
-        GenomeVariant variant = new GenomeVariant(new GenomePosition(refDict, Strand.FWD, refDict.getContigNameToID().get(builder.getChrom()), builder.getPos(), PositionType.ONE_BASED), builder.getRef(), record.get("alt"));
+        GenomeVariant variant = new GenomeVariant(new GenomePosition(refDict, Strand.FWD, refDict.getContigNameToID().get(builder.getChrom()), (int) builder.getPos(), PositionType.ONE_BASED), builder.getRef(), record.get("alt"));
         builder.getHgvs().add(variant.toString());
 
         //Add scores and predictions
